refactor: Profile의 패키지를 도메인에서 쿼리로 변경
