+++ src/main/java/daybyquest/user/application/GetMyProfileService.java
@@ -1,8 +1,8 @@
 package daybyquest.user.application;
 
-import daybyquest.user.domain.Profile;
 import daybyquest.user.domain.UserImages;
 import daybyquest.user.dto.response.MyProfileResponse;
+import daybyquest.user.query.Profile;
 import daybyquest.user.query.ProfileDao;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
