+++ src/main/java/daybyquest/user/query/Profile.java
@@ -1,4 +1,4 @@
-package daybyquest.user.domain;
+package daybyquest.user.query;
 
 import lombok.Getter;
